requirements.txt
umap-learn
numpy
numba
matplotlib
seaborn
scikit-learn
tqdm
networkx
